etl/requirements.txt
psycopg2-binary
pandas
sqlalchemy
python-dotenv
requests
pyyaml
python-slugify
boto3
